move detection logic to classifier
