+++ rrs/misinformation_detection/main.py
@@ -3,7 +3,7 @@ import asyncio
 import logging
 import os
 from datetime import date, datetime, timedelta, timezone
-from typing import Literal, Optional
+from typing import Optional
 from urllib.parse import quote
 
 import duckdb
@@ -12,9 +12,8 @@ import psycopg
 from dotenv import load_dotenv
 
 from mistralai.client import Mistral
-from mistralai.client.types import BaseModel
-from mistralai.extra.run.context import RunContext
 from rrs.dictionary.upsert_subjects import subject_id as make_subject_id
+from rrs.misinformation_detection.classifier import PRICING, MisinfoResult, classify_one
 from rrs.misinformation_detection.definitions import build_system_prompt
 from rrs.utils.generate_id import get_consistent_hash
 
@@ -155,48 +154,6 @@ def _enrich_with_plaintext(segments: pd.DataFrame) -> pd.DataFrame:
 
 # ── LLM classification ────────────────────────────────────────────────────────
 
-PRICING = {
-    "mistral-small-2603": (0.15, 0.60),
-    "mistral-small-3.2": (0.10, 0.30),
-    "mistral-medium-2505": (0.40, 2.00),
-    "mistral-large-2411": (2.00, 6.00),
-}
-
-
-class MisinfoResult(BaseModel):
-    label: Literal["oui", "non", "incertain"]
-    score: float
-    justification: str
-
-
-async def _classify_one(
-    client: Mistral,
-    semaphore: asyncio.Semaphore,
-    system_prompt: str,
-    index: int,
-    total: int,
-    text: str,
-    model: str,
-) -> tuple[MisinfoResult, int, int]:
-    async with semaphore:
-        logging.info(f"[{index + 1}/{total}] Classifying...")
-        async with RunContext(model=model, output_format=MisinfoResult) as run_ctx:
-            run_result = await client.beta.conversations.run_async(
-                run_ctx=run_ctx,
-                instructions=system_prompt,
-                inputs=[
-                    {
-                        "role": "user",
-                        "content": f"Analyse cet extrait et détecte une éventuelle désinformation :\n\n{text}",
-                    }
-                ],
-            )
-        output_text = run_result.output_entries[0].content if run_result.output_entries else ""
-        input_tokens = len(system_prompt + text) // 4
-        output_tokens = len(output_text) // 4
-        return run_result.output_as_model, input_tokens, output_tokens
-
-
 async def classify_segments(
     df: pd.DataFrame,
     subject: str,
@@ -216,7 +173,7 @@ async def classify_segments(
 
     texts = df["plaintext"].tolist()
     tasks = [
-        _classify_one(client, semaphore, system_prompt, i, len(df), str(t), model)
+        classify_one(client, semaphore, system_prompt, i, len(df), str(t), model)
         if not (pd.isna(t) or str(t).strip() == "")
         else _skip()
         for i, t in enumerate(texts)
